adding function which combines logic for checking permissions based on profile-level permissions and role-level permissions.
if profile_permissions has a row describing a permission, it takes precedence and the query should be allowed or
disallowed according to what it specifies. if there is no relevant profile_permission, check role permissions.

because of this, we needed more information than just boolean to decide if a query should be allowed or not because of profile-level permissions